src/backend/requirements.txt
# Core LangChain and LangGraph
langchain==0.3.13
langgraph==0.2.60
langchain-core==0.3.28
langchain-community==0.3.13

# Google Gemini Integration
langchain-google-genai==2.0.8
google-generativeai==0.8.3

# FastAPI and Web Framework
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
pydantic==2.10.4
pydantic-settings==2.7.0

# Web Scraping
requests==2.32.3
beautifulsoup4==4.12.3
lxml==5.3.0
httpx==0.28.1

# Vector Store and Memory
chromadb==0.5.23
faiss-cpu==1.9.0.post1
sentence-transformers==3.3.1

# LangSmith Observability
langsmith==0.2.7

# PDF Generation
reportlab==4.2.5
pypdf2==3.0.1
weasyprint==63.1

# Visualization
matplotlib==3.9.4
plotly==5.24.1
kaleido==0.2.1
pandas==2.2.3
numpy==1.26.4

# Utilities
python-dotenv==1.0.1
tiktoken==0.8.0
tenacity==9.0.0
aiofiles==24.1.0

# Development
pytest==8.3.4
black==24.10.0

